+++ FHIR-Converter/client/src/core/common/enum/metadata-type.ts
@@ -1,0 +1,9 @@
+/*!
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License. See License in the project root for license information.
+ */
+
+export enum MetadataType { 
+	ccda = 'ccda', 
+	hl7v2 = 'hl7v2' 
+}